Revert "[refactor] 페이지 이름 소문자로 변경"
