+++ src/Pages/MainPage.tsx
@@ -1,25 +1,19 @@
 import React, { useState } from 'react'
 import { motion } from 'framer-motion'
-import SideMenuBar from '../components/mainPageComponents/sideMenu'
-import Group from '../components/mainPageComponents/groups'
-import Search from '../components/mainPageComponents/search'
-import Typing from '../components/mainPageComponents/typing'
-import D3Canvas from '../components/mainPageComponents/d3'
-import LocationButton from '../components/mainPageComponents/locationButton'
+import SideMenuBar from '../components/MainPageComponents/SideMenu'
+import Group from '../components/MainPageComponents/Groups'
+import Search from '../components/MainPageComponents/Search'
+import Typing from '../components/MainPageComponents/Typing'
+import D3Canvas from '../components/MainPageComponents/D3'
+import LocationButton from '../components/MainPageComponents/LocationButton'
 
-type Log = {
-  createdAt: string
-  name: string
-  content: string
-}
 
 const MainPage: React.FC = () => {
   const [isCollapsed, setIsCollapsed] = useState(false)
-  const [logs, setLogs] = useState<Log[]>([])
+  const [logs, setLogs] = useState<string[]>([])
 
-  const addLog = (log: Log) => {
+  const addLog = (log: string) => {
     setLogs((prevLogs) => [...prevLogs, log])
-    console.log('Log added:', log) // Optional: 로그가 추가되었음을 확인하기 위한 콘솔 출력
   }
 
   const moveToUserNode = () => {
