+++ src/components/MainPageComponents/SideMenu.tsx
@@ -2,8 +2,8 @@ import React, { Dispatch, SetStateAction } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { motion, AnimatePresence } from 'framer-motion'
 import Logo from '../../assets/Logo.png'
-import Log from '../sideMenuComponents/Log'
-import LoginStatus from '../../modal/userModal/LoginStatus'
+import Log from '../../components/SideMenuComponents/Log'
+import LoginStatus from '../../modal/UserModal/LoginStatus'
 
 type LogItem = {
   content: string
